+++ sqlite3/CHANGELOG.md
@@ -3,6 +3,13 @@
 - Add support for application-defined window functions. To register a custom
   window function, implement `WindowFunction` and register your function with
   `database.registerAggregateFunction`.
+- __Breaking__ (For the experimental `package:sqlite3/wasm.dart` library):
+  - The IndexedDB implementation now stores data in 4k blocks instead of full files.s
+  - Removed `IndexedDbFileSystem.load`. Use `IndexedDbFileSystem.open` instead.
+  - An `IndexedDbFileSystem` now stores all files, the concept of a persistence
+    root has been removed.
+    To access independent databases, use two `IndexedDbFileSystem`s with a different
+    database name.
 
 ## 1.6.4